+++ src/components/main/heatmap/Heatmap.scss
@@ -97,7 +97,6 @@
       overflow: hidden;
       opacity: 0;
       transition: max-height 0.3s ease, opacity 0.3s ease;
-      padding-top: 8px;
 
       li {
         list-style: none;
@@ -109,8 +108,8 @@
           font-size: 0.875rem;
           color: #333;
           text-decoration: none;
-          padding-bottom: 8px;
-
+          padding-bottom: 6px;
+          padding-top: 6px;
           &:hover {
             background-color: #f5f5f5;
           }
@@ -120,7 +119,6 @@
       &.show {
         max-height: 300px;
         opacity: 1;
-        padding-top: 8px;
       }
     }
   }
