+++ src/components/main/heatmap/Heatmap.tsx
@@ -1,175 +1,40 @@
 'use client';
 
-import CalHeatmap from 'cal-heatmap';
-import { useQuery } from '@tanstack/react-query';
-import { useState, useEffect, useRef } from 'react';
-import Tooltip from 'cal-heatmap/plugins/Tooltip';
-import LegendLite from 'cal-heatmap/plugins/LegendLite';
-import CalendarLabel from 'cal-heatmap/plugins/CalendarLabel';
-import type { PluginDefinition } from 'cal-heatmap';
-import { useFetch } from '@/hooks/useFetch';
-import useGetAccessToken from '@/hooks/useGetAccessToken';
-import { format } from 'date-fns';
-import useCheckAccess from '@/hooks/useCheckExistAccess';
-import { usePrevDomainStep } from '@/hooks/usePrevDomainStep';
-import { useNextDomainStep } from '@/hooks/useNextDomainStep';
-import 'cal-heatmap/cal-heatmap.css';
+import { useState } from 'react';
+import { useHeatmapController } from '@/hooks/main/heatmap/useHeatmapController';
+import { useHeatmapData } from '@/hooks/main/heatmap/useHeatmapData';
+import { useHeatmapYearDropdown } from '@/hooks/main/heatmap/useHeatmapYearDropdown';
+import { usePrevDomainStep } from '@/hooks/main/heatmap/usePrevDomainStep';
+import { useNextDomainStep } from '@/hooks/main/heatmap/useNextDomainStep';
 import './Heatmap.scss';
+import 'cal-heatmap/cal-heatmap.css';
 
-const weekdays = ['Sun', 'Mon', 'Tues', 'Wed', 'Thur', 'Fri', 'Sat'];
 const currentDate = new Date();
 const basicYear = currentDate.getFullYear();
-
-const rawCurrentMonth = currentDate.getMonth() + 1; // 0 ~ 11
+const rawCurrentMonth = currentDate.getMonth() + 1;
 const adjustedCurrentMonth = rawCurrentMonth >= 9 ? 9 : rawCurrentMonth;
 
-interface TilApiResponse {
-  data: {
-    year: number;
-    tils: Record<string, number[]>;
-  };
-}
-
-const monthMap: Record<string, string> = {
-  jan: '01', feb: '02', mar: '03', apr: '04',
-  may: '05', jun: '06', jul: '07', aug: '08',
-  sep: '09', oct: '10', nov: '11', dec: '12',
-};
-
-interface TilData {
-  date: string;
-  count: number;
-}
-
-
-
 const Heatmap = () => {
-  const [cal, setCal] = useState<CalHeatmap | null>(null);
   const [year, setYear] = useState(basicYear);
   const [currentMonth, setCurrentMonth] = useState(adjustedCurrentMonth);
-  const heatmapRef = useRef<CalHeatmap | null>(null);
-  const [isOpen, setIsOpen] = useState(false);
-  const { callApi } = useFetch();
-  const accessToken = useGetAccessToken();
-  const existAccess = useCheckAccess(accessToken);
+
+  const { data: tilData = [] } = useHeatmapData(year);
+  const { cal } = useHeatmapController(tilData, year, currentMonth);
+  const { isOpen, setIsOpen, handleNextClick } = useHeatmapYearDropdown(
+    setYear, setCurrentMonth, cal
+  );
+
   const prevStep = usePrevDomainStep(currentMonth);
   const nextStep = useNextDomainStep(currentMonth);
-  
-  const { data: tilData = [] } = useQuery<TilData[]>({
-    queryKey: ['til-data'] as const,
-    queryFn: async () => {
-      const res = await callApi<TilApiResponse>({
-        method: 'GET',
-        endpoint: `/users/tils?year=${year}`,
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
-        credentials: 'include',
-      });
-
-      const raw = res.data?.tils ?? {};
-      const formatted = Object.entries(raw).flatMap(([month, counts]) =>
-        counts.map((count, index) => {
-          const day = String(index + 1).padStart(2, '0');
-          return {
-            date: `${year}-${monthMap[month]}-${day}`,
-            count,
-          };
-        })
-      );
-
-      return formatted;
-    },
-    enabled: existAccess,
-    staleTime: 600000,
-    gcTime: 3600000,
-    //til 생성 후 수동 갱신
-  });
-
-  useEffect(() => {
-    if (!heatmapRef.current && tilData.length > 0) {
-      const startMonthDate = new Date(`${year}-${String(currentMonth).padStart(2, '0')}-01`);
-
-      heatmapRef.current = new CalHeatmap();
-
-      const plugins: PluginDefinition[] = [
-        [
-          Tooltip,
-          {
-            text: (date, value) =>
-              `${value ?? 'No'} contributions on </br>${format(date, 'EEEE, MMMM d, yyyy')}`,
-          },
-        ],
-        [
-          LegendLite,
-          {
-            includeBlank: true,
-            itemSelector: '#ex-ghDay-legend',
-            radius: 2,
-            width: 11,
-            height: 11,
-            gutter: 4,
-          },
-        ],
-        [
-          CalendarLabel,
-          {
-            width: 30,
-            textAlign: 'start',
-            text: () => weekdays.map((d, i) => (i % 2 === 0 ? '' : d)),
-            padding: [25, 0, 0, 0],
-          },
-        ],
-      ];
-
-      heatmapRef.current.paint(
-        {
-          data: {
-            source: tilData,
-            type: 'json',
-            x: 'date',
-            y: 'count',
-            groupY: 'max',
-          },
-          date: { 
-            start: startMonthDate
-          },
-          range: 4,
-          scale: {
-            color: {
-              type: 'linear',
-              range: ['#B3E7FC', '#5FC9F8', '#00B6F9', '#0075A8', '#004B70'],
-              domain: [1, 2, 3, 4, 5],
-            },
-          },
-          domain: {
-            type: 'month',
-            label: { text: 'MMM', textAlign: 'start', position: 'top' },
-          },
-          subDomain: { type: 'ghDay', radius: 2, width: 11, height: 11, gutter: 4 },
-          itemSelector: '#ex-ghDay',
-        },
-        plugins,
-      );
-
-      setCal(heatmapRef.current);
-    }
-  }, [cal, year, currentMonth, tilData]);
-
-  const handleNextClick = (year: number) => {
-    cal?.jumpTo(`${year}-01-30`, true);
-  };
 
   const handlePrevDomain = () => {
     if (prevStep === 0) return;
-  
     setCurrentMonth(prev => prev - prevStep);
     cal?.previous(prevStep);
   };
-  
+
   const handleNextDomain = () => {
     if (nextStep === 0) return;
-  
     setCurrentMonth(prev => prev + nextStep);
     cal?.next(nextStep);
   };
@@ -179,13 +44,13 @@ const Heatmap = () => {
       <div className="heatmap-wrapper__container">
         <div className="heatmap-wrapper__content">
           <div className="heatmap-wrapper__controls-left">
-            <button  className={prevStep === 0 ? 'disabled' : ''} onClick={handlePrevDomain}>&lt;</button>
+            <button className={prevStep === 0 ? 'disabled' : ''} onClick={handlePrevDomain}>&lt;</button>
           </div>
 
           <div id="ex-ghDay" className="heatmap-wrapper__calendar"></div>
 
           <div className="heatmap-wrapper__controls-right">
-            <button className={nextStep === 0 ? 'disabled' : ''} onClick={handleNextDomain} >&gt;</button>
+            <button className={nextStep === 0 ? 'disabled' : ''} onClick={handleNextDomain}>&gt;</button>
           </div>
         </div>
 
@@ -208,14 +73,7 @@ const Heatmap = () => {
         <div className={`heatmap-wrapper__dropdown-content ${isOpen ? 'show' : ''}`}>
           {[0, 1, 2, 3, 4].map(offset => (
             <li key={offset}>
-              <a
-                onClick={() => {
-                  setIsOpen(false);
-                  handleNextClick(basicYear + offset);
-                  setYear(basicYear + offset);
-                  setCurrentMonth(1);
-                }}
-              >
+              <a onClick={() => handleNextClick(basicYear + offset)}>
                 {basicYear + offset}
               </a>
             </li>
