+++ src/hooks/main/heatmap/useNextDomainStep.tsx
